requirements.txt
flask
requests
python-dotenv
PyPDF2
google-generativeai
